+++ hardhat/contracts/GlobalPause.sol
@@ -7,15 +7,15 @@ import {PausableUpgradeable} from "@openzeppelin/contracts-upgradeable/security/
 import {GlobalOwnableUpgradeable} from "./abstracts/GlobalOwnableUpgradeable.sol";
 
 /**
- * @title GlobalPauser
+ * @title GlobalPause
  * @author Lila Rest (lila@ledgity.com)
  * @notice This contract is used to maintain a pause state shared by all contracts of this
  * codebase. It is then read by all contracts that inherit from GlobalPausableUpgradeable 
  * abstract contract to restrict execution of some functions when this contract is not paused.
- * @dev For more details see "GlobalPauser" section of whitepaper.
+ * @dev For more details see "GlobalPause" section of whitepaper.
  * @custom:security-contact security@ledgity.com
  */
-contract GlobalPauser is Initializable, UUPSUpgradeable, GlobalOwnableUpgradeable, PausableUpgradeable {
+contract GlobalPause is Initializable, UUPSUpgradeable, GlobalOwnableUpgradeable, PausableUpgradeable {
     /**
      * @dev Prevents implementation contract from being initialized as recommended by
      * OpenZeppelin.