+++ hardhat/scripts/deploy-GlobalPauser.ts
@@ -1,6 +1,6 @@
 import { deployProxy } from "./lib/deployProxy";
 
-const main = async () => await deployProxy("GlobalPauser", true);
+const main = async () => await deployProxy("GlobalPause", true);
 
 export default main().catch((error) => {
   console.error(error);